phase-3-decoder/requirements.txt
accelerate>=0.34.0
datasets>=3.0.0
huggingface_hub>=0.25.0
torch>=2.3.0
transformers>=4.44.0
tqdm>=4.66.0
pyyaml>=6.0.0
